# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "English Brain API",
    "description": "English Brain Backend - Orchestrator for STT, Ollama LLM, TTS and FSRS Spaced Repetition",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/stats": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Get Stats",
        "operationId": "get_stats_api_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/stats/weekly-report": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "Get Weekly Report",
        "operationId": "get_weekly_report_api_stats_weekly_report_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyReportOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StatsOut": {
        "properties": {
          "total_sessions": {
            "type": "integer",
            "title": "Total Sessions"
          },
          "total_turns": {
            "type": "integer",
            "title": "Total Turns"
          },
          "total_mistakes": {
            "type": "integer",
            "title": "Total Mistakes"
          },
          "cards_due_today": {
            "type": "integer",
            "title": "Cards Due Today"
          },
          "total_cards": {
            "type": "integer",
            "title": "Total Cards"
          },
          "streak_days": {
            "type": "integer",
            "title": "Streak Days"
          },
          "mistakes_by_category": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Mistakes By Category"
          },
          "weekly_activity": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Weekly Activity"
          }
        },
        "type": "object",
        "required": [
          "total_sessions",
          "total_turns",
          "total_mistakes",
          "cards_due_today",
          "total_cards",
          "streak_days",
          "mistakes_by_category",
          "weekly_activity"
        ],
        "title": "StatsOut"
      },
...
      "WeeklyReportOut": {
        "properties": {
          "week_range": {
            "type": "string",
            "title": "Week Range"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "total_sessions": {
            "type": "integer",
            "title": "Total Sessions"
          },
          "total_turns": {
            "type": "integer",
            "title": "Total Turns"
          },
          "total_mistakes": {
            "type": "integer",
            "title": "Total Mistakes"
          },
          "mistakes_this_week": {
            "type": "integer",
            "title": "Mistakes This Week",
            "default": 0
          },
          "mistakes_last_week": {
            "type": "integer",
            "title": "Mistakes Last Week",
            "default": 0
          },
          "markdown_report": {
            "type": "string",
            "title": "Markdown Report"
          }
        },
        "type": "object",
        "required": [
          "week_range",
          "generated_at",
          "total_sessions",
          "total_turns",
          "total_mistakes",
          "markdown_report"
        ],
        "title": "WeeklyReportOut"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}